requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==1.3.0
setuptools==75.6.0
six==1.17.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yt-dlp==2024.12.13
